refactor: 기본 점수 타입 변경

+++ src/main/java/chess/domain/piece/Pawn.java
@@ -11,7 +11,7 @@ import chess.domain.piece.exception.ObstacleException;
 public class Pawn extends AbstractPiece {
 
     private static final double WEEK_SCORE = 0.5;
-    private static final int DEFAULT_SCORE = 1;
+    private static final double DEFAULT_SCORE = 1;
 
     private final List<Direction> forwardDirections = List.of(Direction.UP, Direction.UP_UP);
     private final List<Direction> diagonalDirections = List.of(Direction.UP_LEFT, Direction.UP_RIGHT);
